+++ src/transformers/apply-aliases.js
